scissor has been stripped down; it will be rebuilt. (#57) leveller: #55 adding low shelf / high shelf options, other refinements. leveller: better GUI with enable/disabled et al. biquad filter adding shelving options, some optimizations DecibelsToLinear now uses exp instead of pow; supposedly much faster Maj7Width: better param range of rotation lots of keyboard & focus tweaks #40 no longer resizing to fill parent window, a Renoise fix #58

+++ WaveSabreCore/include/WaveSabreCore/Maj7.hpp
@@ -254,12 +254,12 @@ namespace WaveSabreCore
 #ifdef MIN_SIZE_REL
 #pragma message("Maj7::~Maj7() Leaking memory to save bits.")
 #else
-				for (int i = 0; i < std::size(mpModulations); ++i)
+				for (size_t i = 0; i < std::size(mpModulations); ++i)
 				{
 					delete mpModulations[i];
 				}
 
-				for (int i = 0; i < gModLFOCount; ++i) {
+				for (size_t i = 0; i < gModLFOCount; ++i) {
 					delete mpLFOs[i];
 					delete mpOscillatorDevices[i];
 					delete mpSamplerDevices[i];
